Add files via upload
This is a massive update of the CSS file. Now the website is flexible regards the size of the screen less than 1400px in two modes of landscape and portrait. The font size and the width of the main, header, and nav elements are sensitive to the screen size of the device.  you can check out the end of the CSS file to find out what I mean and you also can read the CSS file.